# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sorteio Ativo",
    "version": "1.0.0",
    "description": "Documentação da API do Sorteio Ativo"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/orders/{id}": {
      "get": {
        "summary": "Obtém os detalhes de um pedido específico (Admin)",
        "description": "(Admin) Retorna os detalhes completos de um pedido, incluindo números associados.",
        "tags": [
          "Admin - Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do pedido."
          },
          "404": {
            "description": "Pedido não encontrado."
          }
        }
      },
      "put": {
        "summary": "Atualiza o status de um pedido (Admin)",
        "description": "(Admin) Permite alterar o status de um pedido. Se um pedido pendente for alterado para 'completed', os números reservados serão marcados como 'sold'. Se alterado para 'failed' ou 'cancelled', serão libertados.",
        "tags": [
          "Admin - Orders"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "pending",
                      "completed",
                      "failed",
                      "cancelled"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Pedido atualizado com sucesso."
          },
          "400": {
            "description": "Status inválido."
          },
          "409": {
            "description": "Conflito, o pedido já foi processado ou está num estado que não permite alteração."
          }
        }
      }
    },
    "/api/admin/orders": {
      "get": {
        "summary": "Lista todos os pedidos do sistema (Admin)",
        "description": "(Admin) Retorna uma lista de todos os pedidos, juntando informações do utilizador e do produto.",
        "tags": [
          "Admin - Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Uma lista de todos os pedidos."
          },
          "401": {
            "description": "Não autenticado."
          },
          "403": {
            "description": "Acesso negado."
          }
        }
      },
      "post": {
        "summary": "Marca uma ordem como completa (Admin)",
        "description": "(Admin) Altera o status de uma ordem para 'completed' e o status dos seus números de 'reserved' para 'sold'. Útil para confirmar pagamentos manuais.",
        "tags": [
          "Admin - Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "orderId"
                ],
                "properties": {
                  "orderId": {
                    "type": "integer",
                    "description": "O ID da ordem a ser marcada como completa."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ordem atualizada com sucesso."
          },
          "400": {
            "description": "ID da ordem não fornecido."
          },
          "404": {
            "description": "Ordem não encontrada."
          },
          "409": {
            "description": "Conflito, a ordem não pode ser marcada como completa (ex: não está pendente)."
          }
        }
      }
    },
...
  },
...
}